code improvments
