+++ pkg/sl1/sl1.go
@@ -20,8 +20,6 @@ package sl1
 
 import (
 	"fmt"
-	"path"
-	"path/filepath"
 
 	"github.com/marco-ostaska/sl1cmd/pkg/sl1/httpcalls"
 )
@@ -58,70 +56,25 @@ func (bInfo *BasicInfo) Load(api string) error {
 	return nil
 }
 
-// LookupDescByURI looks up a description content for a provided URI
-func (bInfo *BasicInfo) LookupDescByURI(uri string) (string, error) {
+//ListByDesc returns the list of index for provided description(s) lookup
+func (bInfo *BasicInfo) ListByDesc(args []string) ([]int, []error) {
+	var found []int
+	var notFound []error
 
-	for _, u := range *bInfo {
-		if uri == u.URI {
-			return u.Description, nil
-		}
-	}
-	return uri, fmt.Errorf("%s: no such uri", uri)
-
-}
-
-// LookupIdxByDesc lookups up index position for provided BasicInfo.Description
-func (bInfo *BasicInfo) LookupIdxByDesc(d string) (int, error) {
-
-	for i, u := range *bInfo {
-		if d == u.Description {
-			return i, nil
-		}
-	}
-	return 0, fmt.Errorf("%s: no such %v", d, path.Base(path.Dir((*bInfo)[0].URI)))
-
-}
-
-// -------------------------------------
-
-// LookupIDbyDesc looks up sl1id for a provided Description
-func (bInfo *BasicInfo) LookupIDbyDesc(s string) (string, error) {
-	i, err := bInfo.LookupIdxByDesc(s)
-	if err != nil {
-		return "", err
-	}
-
-	return path.Base((*bInfo)[i].URI), nil
-}
-
-// Println formats the output in a user friendly style
-func (bInfo *BasicInfo) Println(args []string) {
-
-	if len(args) == 0 {
-		bInfo.printRange()
-	}
-
-	if len(args) > 0 {
-		bInfo.printByDesc(args)
-	}
-}
-
-// printRange formats the output for every BasicInfo range
-func (bInfo *BasicInfo) printRange() {
-	for _, i := range *bInfo {
-		fmt.Printf("sl1id=%s(%s)\n", i.ID, i.Description)
-	}
-}
-
-// printByDesc formats the output searching by Description
-func (bInfo *BasicInfo) printByDesc(args []string) {
 	for _, a := range args {
 		i, err := bInfo.LookupIdxByDesc(a)
 		if err != nil {
-			fmt.Println(err)
+			notFound = append(notFound, err)
 			continue
 		}
-		fmt.Printf("sl1id=%s(%s)\n", filepath.Base(((*bInfo)[i].URI)), ((*bInfo)[i].Description))
+		found = append(found, i)
 	}
 
+	return found, notFound
+
+}
+
+// FmtMsg returns an user friendly output format based on provided index
+func (bInfo *BasicInfo) FmtMsg(i int) string {
+	return fmt.Sprintf("sl1id=%s(%s)", (*bInfo)[i].ID, (*bInfo)[i].Description)
 }
